refactor: api 레이어 정리
